Update gns3-project-delete.py
added specific projects for deletion, example test1
looks for that projects and deletes

+++ gns3-project-delete.py
@@ -4,6 +4,8 @@ BASE_IP = "http://10.48.229."
 GNS3_USER = "gns3"
 GNS3_PW = "gns3"
 
+#list of specific projects to delete
+PROJECTS_TO_DELETE = ["test1"]
 # Read last octets from datastore
 try:
     with open("last_octet", "r") as f:
@@ -25,14 +27,18 @@ for SERVER_URL in SERVER_URLS:
         version = server.get_version()
         print(f"\nConnected to GNS3 server {SERVER_URL} (version: {version})")
 
-        projects = server.get_projects()
+        projects = server.get_projects() #print out all projects existing in the machine
         if not projects:
             print("  No projects to delete.")
         else:
             for project in projects:
-                print(f"  Deleting project '{project['name']}'...")
-                server.delete_project(project_id=project["project_id"])
-            print("  All projects deleted.")
+                if project["name"] in PROJECTS_TO_DELETE:
+                    print(f" Deleting project '{project['name']} '...")
+                    server.delete_project(project_id=project["project_id"])
+                    print("Specific projects have been deleted!")
+                #print(f"  Deleting project '{project['name']}'...")
+                # server.delete_project(project_id=project["project_id"])
+           # print("  All projects deleted.")
 
     except Exception as e:
         print(f"Failed to process {SERVER_URL}: {e}")
